+++ NEWS.md
@@ -3,4 +3,5 @@
 ## Highlights
 * RGB Color Modification
 * Project circles within a matrix
-
+* Quick Mass Column Convert for data.frame
+* Floor and Cap function